Compile fixes for gcc 4.4.1.
git-svn-id: svn://localhost/public/MoSync/trunk@481 6ad7a17e-d552-4a1e-ace4-f958fe14b117

+++ runtimes/cpp/platforms/sdl/ConfigParser.h
@@ -30,7 +30,7 @@ struct Key {
 struct Skin {
 	std::string model;
 	std::string vendor;
-	int screenX, screenY, screenW, screenH;
+	short screenX, screenY, screenW, screenH;
 	std::vector<Key> keys;
 	std::string selectedImage;
 	std::string unselectedImage;
